feat: update function and slight refactor
